+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/api/EntryPoint.kt
@@ -2,9 +2,13 @@ package com.atlassian.performance.tools.virtualusers.api
 
 import com.atlassian.performance.tools.virtualusers.LoadTest
 import com.atlassian.performance.tools.virtualusers.RestUserGenerator
+import com.atlassian.performance.tools.virtualusers.logs.LogConfiguration
 import com.atlassian.performance.tools.virtualusers.logs.LogConfigurationFactory
 import org.apache.logging.log4j.LogManager
+import org.apache.logging.log4j.core.LoggerContext
+import org.apache.logging.log4j.core.config.AbstractConfiguration
 import org.apache.logging.log4j.core.config.ConfigurationFactory.setConfigurationFactory
+import org.apache.logging.log4j.core.config.ConfigurationSource
 
 /**
  * Main entry point for the virtual users load test.
@@ -16,8 +20,10 @@ fun main(args: Array<String>) {
 }
 
 class Application {
+    private val defaultLogConfigFactory = LogConfigurationFactory(LogConfiguration())
+
     fun tryRunning(args: Array<String>) {
-        setConfigurationFactory(LogConfigurationFactory())
+        setConfigurationFactory(defaultLogConfigFactory)
         try {
             run(args)
         } catch (e: Exception) {
@@ -28,12 +34,25 @@ class Application {
     }
 
     private fun run(args: Array<String>) {
-        val options = VirtualUserOptions.Parser().parse(args)
+        val options: VirtualUserOptions = VirtualUserOptions.Parser().parse(args)
+
+        setLoggingConfiguration(options)
+
         @Suppress("DEPRECATION")
         if (options.help) {
             options.printHelp()
         } else {
             LoadTest(options, RestUserGenerator()).run()
         }
+    }
+
+    private fun setLoggingConfiguration(options: VirtualUserOptions) {
+        val logging = options.behavior.logging
+
+        val logConfig = logging.getConstructor().newInstance() as AbstractConfiguration
+        val logConfigurationFactory = LogConfigurationFactory(logConfig)
+        setConfigurationFactory(logConfigurationFactory)
+        val context = LogManager.getContext(false) as LoggerContext
+        context.start(logConfigurationFactory.getConfiguration(null, ConfigurationSource.NULL_SOURCE))
     }
 }